refactor: changed AuthLayout -> AccountSetup and moved Sign up logic into AccountSetup
